refactor id provider tests
